IMB-146 Add dynamic selectable props for get requests
When creating DB queries the API would previously add a set of default props/column names (id, created_at, updated_at) to the query. This causes an error when selecting from a table without those values.

Added a function to stop default props being added to the query when the original table's configured props was already * as they would be selected anyway